Add confirm for delete user account

+++ client-react/src/components/Profile.js
@@ -70,14 +70,23 @@ export default function UserAccount() {
 
   const deleteUser = (id) => {
     const loggedUser = JSON.parse(localStorage.getItem("_auth_state"));
+    console.log(loggedUser)
     id = loggedUser.userId;
     //console.log(loggedUser.userId);
+    let result = window.confirm("Voulez-vous supprimer votre compte ?");
 
-    axios.delete(`http://localhost:3000/api/users/${id}`).then(() => {
-      //console.log("Cliqué");
-      //console.log(`User ${id} supprimé`);
-      logout();
-    });
+    if (result === true && !loggedUser.isAdmin) {
+      localStorage.clear();
+      axios.delete(`http://localhost:3000/api/users/${id}`).then(() => {
+        //console.log("Cliqué");
+        //console.log(`User ${id} supprimé`);
+        logout();
+      });
+      navigate("/", { replace: true });
+      window.location.reload();
+    } else {
+      return;
+    }
   };
 
   return (
